final factor labeling, updated tutorial and README
